add open in grida button to design preview
