+++ app/lib/components/open-in-editor-button.tsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 
 import { PluginSdk } from "@plugin-sdk/app";
-import type { IReflectNodeReference } from "@design-sdk/figma-node";
 import { isAuthenticated } from "@assistant-fp/auth";
 import { useHistory } from "react-router-dom";
 import { Dialog } from "@material-ui/core";
@@ -15,7 +14,7 @@ import {
  */
 export function OpenInEditorButton(props: {
   disabled?: boolean;
-  scene?: IReflectNodeReference;
+  scene?: { id: string };
   framework?: string;
   app?: any;
   button: TOpenButton;
@@ -50,8 +49,8 @@ export function OpenInEditorButton(props: {
     open(
       buildOpenUrlForEditor({
         filekey: filekey,
-        id: props.scene.id,
-        framework: props.framework,
+        id: props?.scene?.id,
+        framework: props?.framework,
       })
     );
     // ..
@@ -102,12 +101,34 @@ function buildOpenUrlForEditor({
   framework,
 }: {
   filekey: string;
-  id: string;
-  framework: string;
+  id?: string;
+  framework?: string;
 }) {
   // local: http://localhost:6626/files/~
   // staging: https://staging-branch-code.grida.co/files/~
   // production: https://code.grida.co/files/~
   // &mode=isolate
-  return `https://code.grida.co/files/${filekey}?node=${id}&framework=${framework}&mode=isolate`;
+
+  // query params - do not provide optional query param if not provided.
+  // - filekey (path)
+  // - node (optional)
+  // - framework (optional)
+  // - mode (optional)
+
+  const queryParams = {};
+
+  if (id) {
+    queryParams["node"] = id;
+    queryParams["mode"] = "isolate";
+  }
+
+  if (framework) {
+    queryParams["framework"] = framework;
+  }
+
+  const queryString = Object.keys(queryParams)
+    .map((key) => `${key}=${queryParams[key]}`)
+    .join("&");
+
+  return `https://code.grida.co/files/${filekey}?${queryString}`;
 }
